[thin_check (rust)] Pull out space map checking routines
