+++ src/pdata/mod.rs
@@ -7,5 +7,6 @@ pub mod btree_merge;
 pub mod btree_leaf_walker;
 pub mod btree_walker;
 pub mod space_map;
+pub mod space_map_checker;
 pub mod unpack;
